Cleaned up question models
