+++ models/Questions/Questions.ts
@@ -1,0 +1,8 @@
+import create from "./createQuestion";
+import remove from "./deleteQuestion";
+import get from "./getQuestionById";
+import update from "./updateQuestion";
+export const createQuestion = create;
+export const deleteQuestion = remove;
+export const getQuestionById = get;
+export const updateQuestion = update;
